refactor: extract form state management to a dedicated hook for improved modularity and code reuse.
Clarify type definitions for inner fields to enhance type safety and maintainability.

Simplify async save and close logic to handle errors more robustly.
