+++ apps/panel/src/components/organisms/bucket-add-field/BucketAddFieldBusiness.tsx
@@ -6,13 +6,18 @@ import {
   useBucketFieldPopups,
   type BucketFieldPopup
 } from "../../../components/molecules/bucket-field-popup/BucketFieldPopupsContext";
-import {FieldKind, type FieldDefinition, type FieldFormState} from "../../../domain/fields/types";
+import {
+  FieldKind,
+  type FieldDefinition,
+  type FieldFormState,
+  type InnerFieldFormState
+} from "../../../domain/fields/types";
 import {
   addInnerField,
   removeInnerField,
   updateInnerField
 } from "../../../domain/fields/inner-fields";
-import {initForm} from "../../../domain/fields";
+import {useFormState} from "./BucketAddFieldHooks";
 
 function isObjectEffectivelyEmpty(obj: Object): boolean {
   if (obj === null || obj === undefined) return true;
@@ -37,36 +42,6 @@ export type FormErrors = {
   multipleSelectionTab?: Record<string, string>;
   innerFields?: string;
 };
-
-function useFormState(
-  field: FieldDefinition,
-  type: FieldKind,
-  isInner: boolean,
-  initialValues?: FieldFormState
-) {
-  const [formValues, setFormValues] = useState<FieldFormState>(field.creationFormDefaultValues);
-  const [formErrors, setFormErrors] = useState<FormErrors>({});
-  const [isInitialized, setIsInitialized] = useState(false);
-
-  useEffect(() => {
-    if (!type) return;
-
-    const base = initForm(type, initialValues);
-
-    const newFormValues = {...base, type};
-    setFormValues(newFormValues);
-    setFormErrors({});
-    setIsInitialized(true);
-  }, [type, initialValues, isInner, field]);
-
-  return {
-    formValues,
-    setFormValues,
-    formErrors,
-    setFormErrors,
-    isInitialized
-  };
-}
 
 const BucketAddFieldBusiness: FC<BucketAddFieldBusinessProps> = ({
   onClose,
@@ -192,7 +167,7 @@ const BucketAddFieldBusiness: FC<BucketAddFieldBusinessProps> = ({
     []
   );
 
-  const handleSaveInnerField = useCallback((values: FieldFormState) => {
+  const handleSaveInnerField = useCallback((values: InnerFieldFormState) => {
     setFormValues(prev => updateInnerField(prev, values));
   }, []);
